Move ProcessAction
