+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/ObsoleteStateDetector.scala
@@ -1,9 +1,8 @@
 package pl.touk.nussknacker.ui.process.deployment
 
-import pl.touk.nussknacker.engine.api.deployment.ProcessState
+import pl.touk.nussknacker.engine.api.deployment.{ProcessAction, ProcessState}
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus
 import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleProcessStateDefinitionManager, SimpleStateStatus}
-import pl.touk.nussknacker.restmodel.processdetails.ProcessAction
 
 object ObsoleteStateDetector {
