+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/helpers/MockFetchingProcessRepository.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.ui.api.helpers
 
 import cats.instances.future._
 import com.typesafe.config.{Config, ConfigFactory}
+import pl.touk.nussknacker.engine.api.deployment.ProcessAction
 import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, VersionId}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.restmodel.displayedgraph.DisplayableProcess
@@ -47,7 +48,7 @@ class MockFetchingProcessRepository(processes: List[BaseProcessDetails[_]])(impl
     getUserProcesses[Unit].map(_.find(p => p.processId == processId).map(_.processingType).get)
 
   //TODO: Implement
-  override def fetchProcessActions(processId: ProcessId)(implicit ec: ExecutionContext): Future[List[processdetails.ProcessAction]] = ???
+  override def fetchProcessActions(processId: ProcessId)(implicit ec: ExecutionContext): Future[List[ProcessAction]] = ???
 
   //TODO: Implement
   override def fetchProcessDetails(processName: ProcessName)(implicit ec: ExecutionContext): Future[Option[ProcessEntityData]] = ???
